Revert "SidebarSettings: Clear all apps after sidebar status is changed"
This reverts commit 68e54998125159f22afd7f2a05ad52327af10cc7.

+++ sidebar/app/src/main/java/com/libremobileos/sidebar/ui/sidebar/SidebarSettingsPage.kt
@@ -62,7 +62,9 @@ fun SidebarSettingsPage(
                     override val onCheckedChange: (Boolean) -> Unit = {
                         mainChecked.value = it
                         viewModel.setSidebarEnabled(it)
-                        viewModel.deleteAllSidebarApps()
+                        if(!it) {
+                            viewModel.deleteAllSidebarApps()
+                        }
                     }
                 })
